+++ apps/graduate/src/pages/admin/schedule/index.ts
@@ -1,0 +1,1 @@
+export { default as ScheduleAdminPage } from './ui/ScheduleAdminPage';
