+++ apps/graduate/src/pages/admin/schedule/styles/ScheduleAdminPage.css.ts
@@ -1,0 +1,9 @@
+import { style } from '@vanilla-extract/css';
+
+import { vars } from '~/vars.css';
+
+export const container = style({
+  padding: vars.spacing.xl,
+  maxWidth: '1200px',
+  margin: '0 auto',
+});
