+++ apps/graduate/src/pages/admin/schedule/styles/ScheduleDescription.css.ts
