+++ apps/graduate/src/pages/admin/schedule/types/schedule.ts
